+++ test/cache.test.js
@@ -1,7 +1,5 @@
 import { applyMiddleware, combineReducers, createStore } from 'redux'
 import thunk from 'redux-thunk'
-// eslint-disable-next-line no-unused-vars
-import regeneratorRuntime from 'regenerator-runtime'
 
 import ApiUnrest from '../src'
 import { timeout } from './utils'
@@ -26,8 +24,8 @@ describe('Api unrest provides a cache', () => {
                   'Content-Type': 'application/json',
                 }[key]),
             },
-            // eslint-disable-next-line require-await
-            json: async () => ({
+
+            json: () => ({
               objects: [i++],
             }),
           }
@@ -81,8 +79,8 @@ describe('Api unrest provides a cache', () => {
                   'Content-Type': 'application/json',
                 }[key]),
             },
-            // eslint-disable-next-line require-await
-            json: async () => ({
+
+            json: () => ({
               objects: [i++],
             }),
           }
@@ -138,8 +136,8 @@ describe('Api unrest provides a cache', () => {
                   'Content-Type': 'application/json',
                 }[key]),
             },
-            // eslint-disable-next-line require-await
-            json: async () => ({
+
+            json: () => ({
               objects: [i++],
             }),
           }
@@ -194,8 +192,8 @@ describe('Api unrest provides a cache', () => {
                   'Content-Type': 'application/json',
                 }[key]),
             },
-            // eslint-disable-next-line require-await
-            json: async () => ({
+
+            json: () => ({
               objects: [i++],
             }),
           }
@@ -246,8 +244,8 @@ describe('Api unrest provides a cache', () => {
                     'Content-Type': 'application/json',
                   }[key]),
               },
-              // eslint-disable-next-line require-await
-              json: async () => ({
+
+              json: () => ({
                 objects: [i++],
               }),
             }
@@ -292,8 +290,8 @@ describe('Api unrest provides a cache', () => {
                   'Content-Type': 'application/json',
                 }[key]),
             },
-            // eslint-disable-next-line require-await
-            json: async () => ({
+
+            json: () => ({
               objects: [url],
             }),
           }
@@ -341,8 +339,8 @@ describe('Api unrest provides a cache', () => {
                   'Content-Type': 'application/json',
                 }[key]),
             },
-            // eslint-disable-next-line require-await
-            json: async () => ({
+
+            json: () => ({
               objects: [url],
             }),
           }
@@ -390,8 +388,8 @@ describe('Api unrest provides a cache', () => {
                   'Content-Type': 'application/json',
                 }[key]),
             },
-            // eslint-disable-next-line require-await
-            json: async () => ({
+
+            json: () => ({
               objects: [url],
             }),
           }
@@ -443,8 +441,8 @@ describe('Api unrest provides a cache', () => {
                   'Content-Type': 'application/json',
                 }[key]),
             },
-            // eslint-disable-next-line require-await
-            json: async () => ({
+
+            json: () => ({
               objects: [url],
             }),
           }